Unique contraints in schema and error handling for soup type and tag

+++ recipe-api/src/soup-type/soup-type.service.ts
@@ -15,25 +15,57 @@ export class SoupTypeService {
   }
 
   async getAllSoupType(): Promise<SoupType[] | null> {
-    return this.prisma.soupType.findMany();
+    return this.prisma.soupType.findMany({
+      orderBy: [{ typeName: 'asc' }],
+    });
   }
 
   async createSoupType(typeName: string): Promise<SoupType> {
-    return this.prisma.soupType.create({
-      data: { typeName: typeName },
-    });
+    if (await this.checkNameExists(typeName)) {
+      throw new Error(
+        'createSoupType: soup type with that name already exists!',
+      );
+    } else {
+      return this.prisma.soupType.create({
+        data: { typeName: typeName },
+      });
+    }
   }
 
   async editSoupType(typeId: number, typeName: string): Promise<SoupType> {
-    return this.prisma.soupType.update({
-      where: { typeId: typeId },
-      data: { typeName: typeName },
-    });
+    if (await this.checkNameExists(typeName)) {
+      throw new Error('editSoupType: soup type with that name already exists!');
+    } else if (await !this.checkIdExists(typeId)) {
+      throw new Error('editSoupType: soup type id does not exist!');
+    } else {
+      return this.prisma.soupType.update({
+        where: { typeId: typeId },
+        data: { typeName: typeName },
+      });
+    }
   }
 
   async deleteSoupType(typeId: number): Promise<SoupType> {
-    return this.prisma.soupType.delete({
-      where: { typeId: typeId },
+    if (await !this.checkIdExists(typeId)) {
+      throw new Error('deleteSoupType: soup type id does not exist!');
+    } else {
+      return this.prisma.soupType.delete({
+        where: { typeId: typeId },
+      });
+    }
+  }
+
+  async checkNameExists(typeName: string): Promise<boolean> {
+    const soupType: SoupType | null = await this.prisma.soupType.findUnique({
+      where: { typeName },
     });
+    return !!soupType;
+  }
+
+  async checkIdExists(typeId: number): Promise<boolean> {
+    const soupType: SoupType | null = await this.prisma.soupType.findUnique({
+      where: { typeId },
+    });
+    return !!soupType;
   }
 }
